21 | Add a "Wrap Paragraph" action
